backend/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Redis
redis==5.0.1
hiredis==2.2.3

# ChromaDB for vector storage
chromadb==0.4.18
sentence-transformers==2.2.2

# Google Gemini
google-generativeai==0.3.1
google-cloud-aiplatform==1.38.1

# Computer Vision
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.26.2

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1

# Async support
aioredis==2.0.1
asyncpg==0.29.0

# Monitoring and logging
loguru==0.7.2
prometheus-client==0.19.0
